Move proxystore.timer to proxystore.utils.timer
